added a delete popup and back button

+++ src/components/pages/Home.js
@@ -8,15 +8,20 @@ const Home = () => {
   useEffect(() => {
     loadUsers();
   }, []); 
-
+ 
   const loadUsers = async () => {
     const result = await axios.get("http://localhost:3030/users");
     setUser(result.data.reverse());
   };
 
   const deleteUser = async id => {
-    await axios.delete(`http://localhost:3030/users/${id}`);
-    loadUsers();
+    const confirm = window.confirm("Would you like to delete ?");
+    if(confirm) {   
+      await axios.delete(`http://localhost:3030/users/${id}`)
+      .then(res => loadUsers())
+      .catch(err => console.log(err));
+      //loadUsers();
+    }
   };
 
   return (
@@ -24,8 +29,8 @@ const Home = () => {
       <h2 className="mt-4 d-inline-block">TC Upload Page</h2>
       <Link className="btn btn-outline-success float-right mt-4" to="/users/add-tc">Add TC's</Link>
       <div className="py-4">
-        <table class="table border shadow table-hover">
-          <thead class="thead-dark">
+        <table className="table border shadow table-hover">
+          <thead className="thead-dark">
             <tr>
               <th scope="col">#</th>
               <th scope="col">Name</th>
@@ -36,7 +41,7 @@ const Home = () => {
           </thead>
           <tbody >
             {users.map((user, index) => (
-              <tr>
+              <tr key={index}>
                 <th scope="row">{index + 1}</th>
                 <td>{user.name}</td>
                 <td>{user.username}</td>
@@ -46,15 +51,11 @@ const Home = () => {
                     View
                   </Link>
                   <Link
-                    class="btn btn-outline-primary mr-2"
-                    to={`/users/edit/${user.id}`}
-                  >
+                    class="btn btn-outline-primary mr-2" to={`/users/edit/${user.id}`}>
                     Edit
                   </Link>
                   <Link
-                    class="btn btn-danger"
-                    onClick={() => deleteUser(user.id)}
-                  >
+                    class="btn btn-danger"onClick={() => deleteUser(user.id)}>
                     Delete
                   </Link>
                 </td>
